apply st-anysize-0.8.1 patch
By default, st's window size always snaps to the nearest multiple of the
character size plus a fixed inner border (set with borderpx in
config.h). When the size of st does not perfectly match the space
allocated to it (when using a tiling WM, for example), unsightly gaps
will appear between st and other apps, or between instances of st. In my
case I had a problem with gaps between st in dwm. This patch fixes the
problem.